requirements.txt
python-docx
pydub
requests
azure-cognitiveservices-speech
python-dotenv
